Add server-side support for storing trip and place additions

+++ emission/tests/analysisTests/intakeTests/TestPipelineRealData.py
@@ -48,10 +48,13 @@ import emission.core.wrapper.localdate as ecwl
 import emission.net.usercache.abstract_usercache_handler as enuah
 import emission.analysis.plotting.geojson.geojson_feature_converter as gfc
 import emission.storage.timeseries.tcquery as estt
+import emission.storage.timeseries.abstract_timeseries as esta
 import emission.core.common as ecc
+import emission.core.wrapper.user as ecwu
 
 # Test imports
 import emission.tests.common as etc
+import time
 
 class TestPipelineRealData(unittest.TestCase):
     def setUp(self):
@@ -62,16 +65,20 @@ class TestPipelineRealData(unittest.TestCase):
         logging.info("setUp complete")
 
     def tearDown(self):
-        logging.debug("Clearing related databases for %s" % self.testUUID)
-        # Clear the database only if it is not an evaluation run
-        # A testing run validates that nothing has changed
-        # An evaluation run compares to different algorithm implementations
-        # to determine whether to switch to a new implementation
-        if not hasattr(self, "evaluation") or not self.evaluation:
-            self.clearRelatedDb()
-        if hasattr(self, "analysis_conf_path"):
-            os.remove(self.analysis_conf_path)
-        logging.info("tearDown complete")
+        if os.environ.get("SKIP_TEARDOWN", False):
+            logging.info("SKIP_TEARDOWN = true, not clearing related databases")
+            ecwu.User.registerWithUUID("automated_tests", self.testUUID)
+        else:
+            logging.debug("Clearing related databases for %s" % self.testUUID)
+            # Clear the database only if it is not an evaluation run
+            # A testing run validates that nothing has changed
+            # An evaluation run compares to different algorithm implementations
+            # to determine whether to switch to a new implementation
+            if not hasattr(self, "evaluation") or not self.evaluation:
+                self.clearRelatedDb()
+            if hasattr(self, "analysis_conf_path"):
+                os.remove(self.analysis_conf_path)
+            logging.info("tearDown complete")
 
     def clearRelatedDb(self):
         logging.info("Timeseries delete result %s" % edb.get_timeseries_db().delete_many({"user_id": self.testUUID}).raw_result)
@@ -695,6 +702,184 @@ class TestPipelineRealData(unittest.TestCase):
         start_ld = ecwl.LocalDate({'year': 2016, 'month': 1, 'day': 16})
         cacheKey = "diary/trips-2016-01-16"
         self.standardMatchDataGroundTruth(dataFile, start_ld, cacheKey)
+
+    def compare_composite_objects(self, ct, et):
+        print(f"--------------- Comparing composite trip {ct['_id']} to expected composite trip {et['_id']} -------------------")
+        self.assertEqual(ct['data']['start_ts'], et['data']['start_ts'])
+        self.assertEqual(ct['data']['end_ts'], et['data']['end_ts'])
+        if 'end_confirmed_place' in et['data']:
+            self.assertEqual(ct['data']['end_confirmed_place']['data']['enter_ts'],
+                                et['data']['end_confirmed_place']['data']['enter_ts'])
+            if 'exit_ts' in et['data']['end_confirmed_place']:
+                self.assertEqual(ct['data']['end_confirmed_place']['exit_ts'],
+                                    et['data']['end_confirmed_place']['exit_ts'])
+        self.assertEqual(len(ct['data']['locations']), len(et['data']['locations']))
+
+    def testJackUntrackedTimeMar12(self):
+        dataFile = "emission/tests/data/real_examples/jack_untracked_time_2023-03-12"
+        etc.setupRealExample(self, dataFile)
+        etc.runIntakePipeline(self.testUUID)
+        ts = esta.TimeSeries.get_time_series(self.testUUID)
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile+".expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
+
+    def testShankariNotUntrackedTimeMar21(self):
+        # https://github.com/e-mission/e-mission-docs/issues/870
+        # This data *used to* process with untracked time.
+        # We tweaked the threshold for untracked time, so from now on this data
+        # should process smoothly into a continuous sequence of confirmed trips.
+        # https://github.com/e-mission/e-mission-server/commit/df9d9f0844eedcf7405d88afe9da1b02ee365986
+        dataFile = "emission/tests/data/real_examples/shankari_not_untracked_time_mar_21"
+        start_run = time.time()
+        etc.setupRealExample(self, dataFile)
+        etc.runIntakePipeline(self.testUUID)
+        end_run = time.time()
+        ts = esta.TimeSeries.get_time_series(self.testUUID)
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        for ct in composite_trips:
+            # for this data, every composite trip should come from a confirmed trip,
+            # NOT from untracked time
+            self.assertEqual(ct['metadata']['origin_key'], 'analysis/confirmed_trip')
+            self.assertGreater(ct["metadata"]["write_ts"], start_run)
+            self.assertLessEqual(ct["metadata"]["write_ts"], end_run)
+
+    def testShankariNotUntrackedTimeJan15(self):
+        # This data has a reboot, so it should process with 1 instance of untracked time
+        dataFile = "emission/tests/data/real_examples/shankari_untracked_time_jan_15_reboot_multi_day"
+        start_run = time.time()
+        etc.setupRealExample(self, dataFile)
+        etc.runIntakePipeline(self.testUUID)
+        end_run = time.time()
+        ts = esta.TimeSeries.get_time_series(self.testUUID)
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        countUntrackedTime = 0
+        for ct in composite_trips:
+            if ct['metadata']['origin_key'] == 'analysis/confirmed_untracked':
+                countUntrackedTime += 1
+            self.assertGreater(ct["metadata"]["write_ts"], start_run)
+            self.assertLessEqual(ct["metadata"]["write_ts"], end_run)
+        self.assertEqual(countUntrackedTime, 1)
+
+    def testShankariUntrackedTimeJul20(self):
+        # This data has a large gap, so it should process with 1 instance of untracked time
+        dataFile = "emission/tests/data/real_examples/shankari_untracked_time_jul_20_large_gap"
+        start_run = time.time()
+        etc.setupRealExample(self, dataFile)
+        etc.runIntakePipeline(self.testUUID)
+        end_run = time.time()
+        ts = esta.TimeSeries.get_time_series(self.testUUID)
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        countUntrackedTime = 0
+        for ct in composite_trips:
+            logging.debug("composite trip metadata %s = " % ct['metadata'])
+            if ct['metadata']['origin_key'] == 'analysis/confirmed_untracked':
+                countUntrackedTime += 1
+            self.assertGreater(ct["metadata"]["write_ts"], start_run)
+            self.assertLessEqual(ct["metadata"]["write_ts"], end_run)
+        self.assertEqual(countUntrackedTime, 0)
+
+    def testCompositeTripIncremental(self):
+        # Test for 545114feb5ac15caac4110d39935612525954b71
+        dataFile_1 = "emission/tests/data/real_examples/shankari_2016-08-04"
+        dataFile_2 = "emission/tests/data/real_examples/shankari_2016-08-05"
+
+        etc.setupRealExample(self, dataFile_1)
+        etc.runIntakePipeline(self.testUUID)
+
+        ts = esta.TimeSeries.get_time_series(self.testUUID)
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile_1+".before-user-inputs.expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
+
+        self.entries = json.load(open(dataFile_2+".user_inputs"), object_hook = bju.object_hook)
+        # Load the place additions from the 5th (so after the end of the current day)
+        etc.setupRealExampleWithEntries(self)
+        etc.runIntakePipeline(self.testUUID)
+        # They should all match the final place
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile_1+".all-match-last-place.expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
+
+        # load day 2
+        with open(dataFile_2) as df2:
+            self.entries = json.load(df2, object_hook = bju.object_hook)
+        etc.setupRealExampleWithEntries(self)
+        etc.runIntakePipeline(self.testUUID)
+        # The place additions should be dispersed to the actual places
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile_1+".spread-across-aug-5.expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
+
+        # load place and trip additions and trip inputs for the first day
+        self.entries = json.load(open(dataFile_1+".user_inputs"), object_hook = bju.object_hook)
+        # Load the place additions from the 4th (so somewhere in the middle, and all mixed)
+        etc.setupRealExampleWithEntries(self)
+        etc.runIntakePipeline(self.testUUID)
+        # They should all match the actual entries
+        # Trip matches should also work
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile_1+".trip-matches-check-aug-4.expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
+
+    def testCompositeTripIncrementalLastPlaceMatches(self):
+        # Test for 545114feb5ac15caac4110d39935612525954b71
+        dataFile_1 = "emission/tests/data/real_examples/shankari_2023-04-13"
+        dataFile_2 = "emission/tests/data/real_examples/shankari_2023-04-14"
+
+        etc.setupRealExample(self, dataFile_1)
+        etc.runIntakePipeline(self.testUUID)
+
+        ts = esta.TimeSeries.get_time_series(self.testUUID)
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile_1+".before-user-inputs.expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
+
+        # Load the place additions for both days
+        self.entries = json.load(open(dataFile_1+".user_inputs"), object_hook = bju.object_hook)
+        etc.setupRealExampleWithEntries(self)
+        self.entries = json.load(open(dataFile_2+".user_inputs"), object_hook = bju.object_hook)
+        etc.setupRealExampleWithEntries(self)
+        etc.runIntakePipeline(self.testUUID)
+
+        # They should all match the final place
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile_1+".all-match-last-place.expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
+
+        # load day 2
+        with open(dataFile_2) as df2:
+            self.entries = json.load(df2, object_hook = bju.object_hook)
+        etc.setupRealExampleWithEntries(self)
+        etc.runIntakePipeline(self.testUUID)
+        # The place additions should be dispersed to the actual places
+        composite_trips = list(ts.find_entries(["analysis/composite_trip"], None))
+        with open(dataFile_1+".retained-last-place.expected_composite_trips") as expectation:
+            expected_trips = json.load(expectation, object_hook = bju.object_hook)
+            self.assertEqual(len(composite_trips), len(expected_trips))
+            for i in range(len(composite_trips)):
+                self.compare_composite_objects(composite_trips[i], expected_trips[i])
 
 if __name__ == '__main__':
     etc.configLogging()
